Make not reading the default as long as it is running on .NET 8
